Horrible hack working around my failure to fix a bug properly
At the parent commit da473302, some lines are being incorrectly
truncated at word boundaries. It's something to do with an interaction
between (a) the changes made in da473302 in order to implement
--retain-plus-minus-markers, and (b) the logic in tokenize(). This
commit works around that by reinstating the leading space character
during tokenization and syntax highlighting, and then removing it
afterwards in a hacky way.